requirements.txt
flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
requests==2.32.3
google-cloud-secret-manager==2.21.0
google-cloud-aiplatform==1.71.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
